Refactored, restructured, and cleanup data structure statistics gathering. Doesn't compile yet - neeed to rework schema generation accordingly.

+++ sqml-api/src/main/java/ai/dataeng/sqml/ingest/stats/KeyedSourceRecordStatistics.java
@@ -1,7 +1,9 @@
 package ai.dataeng.sqml.ingest.stats;
 
 import ai.dataeng.sqml.flink.util.FlinkUtilities;
+import ai.dataeng.sqml.ingest.DatasetRegistration;
 import ai.dataeng.sqml.ingest.source.SourceRecord;
+import ai.dataeng.sqml.schema2.basic.ConversionError;
 import org.apache.flink.api.common.state.ValueState;
 import org.apache.flink.api.common.state.ValueStateDescriptor;
 import org.apache.flink.api.common.typeinfo.TypeHint;
@@ -19,40 +21,47 @@ public class KeyedSourceRecordStatistics extends KeyedProcessFunction<Integer, S
 
     private int maxRecords = 100000;
     private int maxTimeInMin = 5;
-    private OutputTag<SourceTableStatistics.Accumulator> statsOutput;
+    private final OutputTag<SourceTableStatistics> statsOutput;
+    private final DatasetRegistration datasetReg;
 
-    private transient ValueState<SourceTableStatistics.Accumulator> stats;
+    private transient ValueState<SourceTableStatistics> stats;
     private transient ValueState<Long> nextTimer;
 
-    public KeyedSourceRecordStatistics(OutputTag<SourceTableStatistics.Accumulator> tag) {
+    public KeyedSourceRecordStatistics(OutputTag<SourceTableStatistics> tag, DatasetRegistration datasetReg) {
         this.statsOutput = tag;
+        this.datasetReg = datasetReg;
     }
 
     @Override
     public void processElement(SourceRecord sourceRecord, Context context, Collector<SourceRecord> out) throws Exception {
-        SourceTableStatistics.Accumulator acc = stats.value();
+        SourceTableStatistics acc = stats.value();
         if (acc == null) {
-            acc = new SourceTableStatistics.Accumulator();
+            acc = new SourceTableStatistics(datasetReg);
             long timer = FlinkUtilities.getCurrentProcessingTime() + TimeUnit.MINUTES.toMillis(maxTimeInMin);
             nextTimer.update(timer);
             context.timerService().registerProcessingTimeTimer(timer);
             //Register an event timer into the far future to trigger when the stream ends
             context.timerService().registerEventTimeTimer(Long.MAX_VALUE);
         }
-        acc.add(sourceRecord);
-        stats.update(acc);
-        if (acc.getCount() >= maxRecords) {
-            context.timerService().deleteProcessingTimeTimer(nextTimer.value());
-            context.output(statsOutput, acc);
-            stats.clear();
-            nextTimer.clear();
+        ConversionError.Bundle<StatsIngestError> errors = acc.validate(sourceRecord);
+        if (errors.isFatal()) {
+            //TODO: Record is flawed, put it in sideoutput and issue warning
+        } else {
+            acc.add(sourceRecord);
+            stats.update(acc);
+            if (acc.getCount() >= maxRecords) {
+                context.timerService().deleteProcessingTimeTimer(nextTimer.value());
+                context.output(statsOutput, acc);
+                stats.clear();
+                nextTimer.clear();
+            }
+            out.collect(sourceRecord);
         }
-        out.collect(sourceRecord);
     }
 
     @Override
     public void onTimer(long timestamp, OnTimerContext ctx, Collector<SourceRecord> out) throws Exception {
-        SourceTableStatistics.Accumulator acc = stats.value();
+        SourceTableStatistics acc = stats.value();
         if (acc != null) ctx.output(statsOutput, acc);
         stats.clear();
         nextTimer.clear();
@@ -60,9 +69,9 @@ public class KeyedSourceRecordStatistics extends KeyedProcessFunction<Integer, S
 
     @Override
     public void open(Configuration config) {
-        ValueStateDescriptor<SourceTableStatistics.Accumulator> statsDesc =
+        ValueStateDescriptor<SourceTableStatistics> statsDesc =
                 new ValueStateDescriptor<>(statsOutput.getId() + STATE_NAME_SUFFIX + ".data",
-                        TypeInformation.of(new TypeHint<SourceTableStatistics.Accumulator>() {
+                        TypeInformation.of(new TypeHint<SourceTableStatistics>() {
                         }));
         stats = getRuntimeContext().getState(statsDesc);
         ValueStateDescriptor<Long> nextTimerDesc =
